nav-footer link fix
For some reason the links in the nav-footer are not clickable. The only fix I have been able to come up with for this is to add a relative positioning to .nav-footer. It does not appear as though there is an overlapping issue or a z-index would have corrected this. This definitely works, though :D